refactor : 패키지 구조 변경
domain과 controller 별로 나누지 않고 UserController와 User는 user 패키지에 저장
