+++ src/main/java/codesquad/qua/QuestionController.java
@@ -1,0 +1,8 @@
+package codesquad.qua;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class QuestionController {
+
+}
