+++ src/main/java/codesquad/user/User.java
@@ -1,4 +1,4 @@
-package codesquad.domain;
+package codesquad.user;
 
 public class User {
