+++ src/main/java/codesquad/user/UserController.java
@@ -1,7 +1,7 @@
-package codesquad.controller;
+package codesquad.user;
 
 
-import codesquad.domain.User;
+import codesquad.user.User;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
